+++ examples/repl.rs
@@ -114,10 +114,15 @@ fn main() {
             }
             "exit" | "quit" => break, // quit
             "scope" => {
-                scope
-                    .iter()
-                    .enumerate()
-                    .for_each(|(i, (name, value))| println!("[{}] {} = {:?}", i + 1, name, value));
+                scope.iter_raw().enumerate().for_each(|(i, (name, value))| {
+                    println!(
+                        "[{}] {}{} = {:?}",
+                        i + 1,
+                        name,
+                        if value.is_shared() { " (shared)" } else { "" },
+                        *value.read_lock::<Dynamic>().unwrap(),
+                    )
+                });
                 continue;
             }
             "astu" => {