+++ RELEASES.md
@@ -25,6 +25,7 @@ New features
 * Capturing of the calling scope for function call via the `func!(...)` syntax.
 * `Module::set_indexer_get_set_fn` is added as a shorthand of both `Module::set_indexer_get_fn` and `Module::set_indexer_set_fn`.
 * New `unicode-xid-ident` feature to allow [Unicode Standard Annex #31](http://www.unicode.org/reports/tr31/) for identifiers.
+* `Scope::iter_raw` returns an iterator with a reference to the underlying `Dynamic` value (which may be shared).
 
 Breaking changes
 ----------------
@@ -33,6 +34,7 @@ Breaking changes
 * Function signature for defining custom syntax is simplified.
 * `Engine::register_raw_fn_XXX` API shortcuts are removed.
 * `PackagesCollection::get_fn`, `PackagesCollection::contains_fn`, `Module::get_fn` and `Module::contains_fn` now take an additional `public_only` parameter indicating whether only public functions are accepted.
+* The iterator returned by `Scope::iter` now contains a clone of the `Dynamic` value (unshared).
 
 Housekeeping
 ------------